+++ lib/src/action.dart
@@ -1,5 +1,6 @@
 library bwu_webdriver.src.io;
 
+import 'dart:math' show Point;
 import 'dart:async' show Future, Stream;
 import 'package:webdriver/io.dart' as wd;
 
@@ -13,22 +14,32 @@ class Action {
   /// Set the [attribute] of [element] to [value].
   Future setAttribute(wd.WebElement element, String attribute, String value) {
     return driver.execute(
-        'arguments[0].setAttribute(arguments[1], arguments[2])', [
-      element,
-      attribute,
-      value
-    ]);
+        'arguments[0].setAttribute(arguments[1], arguments[2])',
+        [element, attribute, value]);
   }
 
-  /// Drag the element found by [sourceElementSelector] onto the element found
-  /// by [targetElementSelector].
+  /// Drag the element found by [sourceSelector] or [sourceLocation] onto
+  /// the element found by [targetSelector] or [targetLocation].
+  /// [sourceSelector] is used as a CSS selector to find the source element.
+  /// [sourceLocation] uses the element found at this position on the page as
+  /// source element.
+  /// [targetSelector] is used as a CSS selector to find the target element.
+  /// [targetLocation] uses the element found at this position on the page as
+  /// source element.
   /// [jQueryUrl] allows to pass an URL for a specific jQuery version or
-  Future dragAndDrop(String sourceElementSelector, String targetElementSelector,
-      {jQueryUrl: _jQueryUrl}) async {
-    await driver.execute(_loadJQueryJs, [jQueryUrl]);
-    String javaScriptString =
-        "${_dragAndDropHelper}\$('${sourceElementSelector}').simulateDragDrop({ dropTarget: '${targetElementSelector}'});";
-    print(javaScriptString);
-    await driver.execute(javaScriptString, []);
+  Future dragAndDrop(
+      {String sourceSelector,
+      Point sourceLocation,
+      String targetSelector,
+      Point targetLocation,
+      String jQueryUrl}) async {
+    await driver.executeAsync(_loadJQueryJs, [jQueryUrl]);
+    await driver.execute(
+        _simulateDragDropScript(
+            sourceSelector: sourceSelector,
+            sourceLocation: sourceLocation,
+            targetSelector: targetSelector,
+            targetLocation: targetLocation),
+        []);
   }
 }