+++ test/all.dart
@@ -1,0 +1,4 @@
+library bwu_webdriver.test.all;
+
+main() {
+}