+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## 0.1.0
-- nothing yet
+- drag and drop using jQuery to work around Chrome
+ChromeDriver issue https://code.google.com/p/chromedriver/issues/detail?id=841